Fix to issue #2
The SD card would sometimes not initialise. A possible fix to this is in this
commit.
By putting a finger on various terminals I found that the CS pin of the SD card adapter
could stop initialisation of the card if it was held high. Holding it low with a high-ish
resistance allowed the SD card to initialise every time. With the theory that the CS line
was activating before or during the SD card initialisation and messing it up, holding it low with a resistor
would deselect it unless the blue pill drove the GPIO high when it wanted ot send a command
or data.
As a test, I have added an explicit GPIO output data direction setup and drive the GPIo to 0 as early
as possible. This does seem to have solved the problem I saw, which hopefully is the problem
seen everywhere.